requirements.txt
flask
flask_sqlalchemy
google-generativeai
sentence-transformers
numpy
scikit-learn
chromadb
